🎨 Refactored next-auth demo
