+++ templates/next-auth/src/components/elements/Navbar/Navbar.tsx
