+++ templates/next-auth/src/components/elements/Navbar/index.ts
