+++ templates/next-auth/src/components/layouts/Default/index.ts
