+++ templates/next-auth/src/components/layouts/index.ts
@@ -1,0 +1,1 @@
+export * from './Default';
