+++ templates/next-auth/src/components/modules/Authentication/index.ts
