+++ templates/next-auth/src/components/modules/Header/index.ts
