+++ templates/next-auth/src/components/modules/WalletInfo/WalletInfo.tsx
