+++ templates/next-auth/src/components/modules/WalletInfo/index.ts
