+++ templates/next-auth/src/components/templates/AuthPage/AuthPage.tsx
@@ -1,4 +1,4 @@
-import { Authentication } from '../../modules';
+import { Authentication } from "src/components/modules";
 
 const AuthPage = () => {
   return <Authentication />;
