+++ templates/next-auth/src/components/templates/AuthPage/index.ts
