+++ templates/next-auth/src/components/templates/UserPage/UserPage.tsx
