+++ templates/next-auth/src/components/templates/UserPage/index.ts
