+++ templates/next-auth/src/components/templates/index.ts
