+++ templates/next-auth/src/constants/time.ts
