+++ templates/next-auth/src/utils/format.ts
